Motion adjustment / refactor / bug fix
