+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/MediaPipe2VrmHand/Scripts/MPHandDebugVisualizer.cs
@@ -4,6 +4,9 @@ using UnityEngine;
 
 namespace Baku.VMagicMirror
 {
+    /// <summary>
+    /// デバッグでBarracudaが吐いてる手の座標情報が見たいときに使うやつ
+    /// </summary>
     public class MPHandDebugVisualizer : MonoBehaviour
     {
         [SerializeField] private MPHand source = null;
