+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/MediaPipe2VrmHand/Scripts/MPHandIK.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 65b91c837ced3be4db99b28bac19d626
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
